requirements.txt
flask
gspread
google-auth
google-api-python-client
